+++ http-client/Network/HTTP/Client/Types.hs
@@ -166,9 +166,12 @@ data HttpExceptionContent
                    -- @since 0.5.0
                    | InvalidHeader S.ByteString
                    -- ^ The given response header line could not be parsed
-                   --   or given request header is not compliant (e.g. has newlines)
                    --
                    -- @since 0.5.0
+                   | InvalidRequestHeader S.ByteString
+                   -- ^ The given request header is not compliant (e.g. has newlines)
+                   --
+                   -- @since 0.5.14
                    | InternalException SomeException
                    -- ^ An exception was raised by an underlying library when
                    -- performing the request. Most often, this is caused by a